requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
geographiclib==1.50
geopy==2.0.0
herepy==3.0.1
idna==2.10
lxml==4.5.2
mysql-connector-python==8.0.21
pip==20.2.2
protobuf==3.13.0
requests==2.24.0
setuptools==49.6.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
urllib3==1.25.10
